requirements.txt
# Production requirements
django==5.2.5
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
whitenoise==6.5.0
django-allauth==0.60.1  # For Google OAuth

# Development requirements
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
pre-commit==3.4.0
django-debug-toolbar==4.2.0
flake8==6.1.0
flake8-docstrings==1.7.0
isort==5.12.0
